Stop placing blocks within yourself
Detect if a block will intersect with a collision entity, if so don't
place it. Also fixed create/destroy cooldowns, and an attempt to fix the
camera jitter at large distances

+++ app/Character/CharacterStats.h
@@ -11,8 +11,8 @@ public:
 
 	int getJumpSpeed() { return _runningSpeed; }
 
-	int getCreateBlockCooldown() { return _createCooldown; }
-	int getDestroyBlockTime() { return _destroyTime; }
+	double getCreateBlockCooldown() { return _createCooldown; }
+	double getDestroyBlockTime() { return _destroyTime; }
 
 	int getManipulateDist() { return _manipulateDistance; }
